+++ src/main/java/kakao/festapick/config/SecurityConfig.java
@@ -11,6 +11,7 @@ import kakao.festapick.jwt.service.JwtService;
 import kakao.festapick.oauth2.handler.SocialSuccessHandler;
 import kakao.festapick.user.domain.UserRoleType;
 import kakao.festapick.user.service.OAuth2UserService;
+import kakao.festapick.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -39,6 +40,7 @@ public class SecurityConfig {
     private final JwtService jwtService;
     private final CookieComponent cookieComponent;
     private final OAuth2UserService oAuth2UserService;
+    private final UserService userService;
     @Value("${spring.front.domain}")
     private String frontDomain;
     @Value("${spring.backend.domain}")
@@ -74,9 +76,9 @@ public class SecurityConfig {
         http.sessionManagement(session->
                 session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
 
-        http.addFilterBefore(new JwtFilter(jwtUtil,oAuth2UserService), LogoutFilter.class);
+        http.addFilterBefore(new JwtFilter(jwtUtil,userService), LogoutFilter.class);
 
-        http.addFilterBefore(new JwtFilterForAdminPage(jwtUtil,oAuth2UserService,cookieComponent), JwtFilter.class);
+        http.addFilterBefore(new JwtFilterForAdminPage(jwtUtil,userService,cookieComponent), JwtFilter.class);
 
         http.addFilterAt(new CustomLogoutFilter(jwtUtil, jwtService, cookieComponent), LogoutFilter.class);
         http.addFilterAfter(new CustomLogoutFilterForAdminPage(jwtUtil, cookieComponent),CustomLogoutFilter.class);
