+++ src/main/java/kakao/festapick/global/filter/JwtFilter.java
@@ -10,6 +10,7 @@ import kakao.festapick.jwt.util.JwtUtil;
 import kakao.festapick.jwt.util.TokenType;
 import kakao.festapick.user.domain.UserEntity;
 import kakao.festapick.user.service.OAuth2UserService;
+import kakao.festapick.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -26,7 +27,7 @@ import java.util.List;
 public class JwtFilter extends OncePerRequestFilter {
 
     private final JwtUtil jwtUtil;
-    private final OAuth2UserService oAuth2UserService;
+    private final UserService userService;
 
     @Override
     protected boolean shouldNotFilter(HttpServletRequest request) throws ServletException {
@@ -61,7 +62,7 @@ public class JwtFilter extends OncePerRequestFilter {
         String identifier = claims.get("identifier").toString();
 
         try {
-            UserEntity findUser = oAuth2UserService.findByIdentifier(identifier);
+            UserEntity findUser = userService.findByIdentifier(identifier);
             List<GrantedAuthority> authorities = Collections.singletonList(new SimpleGrantedAuthority("ROLE_"+findUser.getRoleType().name()));
             Authentication auth = new UsernamePasswordAuthenticationToken(identifier, null, authorities);
             SecurityContextHolder.getContext().setAuthentication(auth);
