+++ src/main/java/kakao/festapick/wish/service/WishService.java
@@ -8,7 +8,7 @@ import kakao.festapick.global.exception.DuplicateEntityException;
 import kakao.festapick.global.exception.ExceptionCode;
 import kakao.festapick.global.exception.NotFoundEntityException;
 import kakao.festapick.user.domain.UserEntity;
-import kakao.festapick.user.service.OAuth2UserService;
+import kakao.festapick.user.service.UserService;
 import kakao.festapick.wish.domain.Wish;
 import kakao.festapick.wish.dto.WishResponseDto;
 import kakao.festapick.wish.repository.WishRepository;
@@ -24,14 +24,14 @@ import org.springframework.transaction.annotation.Transactional;
 public class WishService {
 
     private final WishRepository wishRepository;
-    private final OAuth2UserService oAuth2UserService;
+    private final UserService userService;
     private final FestivalRepository festivalRepository;
 
     @Transactional
     public WishResponseDto createWish(Long festivalId, String identifier) {
         Festival festival = festivalRepository.findFestivalById(festivalId)
                 .orElseThrow(() -> new NotFoundEntityException(ExceptionCode.FESTIVAL_NOT_FOUND));
-        UserEntity user = oAuth2UserService.findByIdentifier(identifier);
+        UserEntity user = userService.findByIdentifier(identifier);
 
         wishRepository.findByUserIdentifierAndFestivalId(identifier, festivalId)
                 .ifPresent(w -> {
